api/requirements.txt
python-multipart
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
pandas
pyodbc
python-dotenv
streamlit
openpyxl
